rename: useCommitteeHeader 훅 이름에 맞게 폴더명 변경
